requirements.txt
Flask==2.0.3
requests==2.27.1
scikit-learn==1.0.2
numpy==1.21.5
tqdm==4.63.0
torch==1.9.0
rich==12.0.0
seqeval==1.2.2
pytorch-crf==0.7.2
gensim==4.3.0
nltk==3.7
